Excepciones, tests y LOGS

+++ BackEnd-I/Clinica-MVC/src/main/java/dh/backend/clinicamvc/service/impl/OdontologoService.java
@@ -1,8 +1,12 @@
 package dh.backend.clinicamvc.service.impl;
 
 import dh.backend.clinicamvc.entity.Odontologo;
+import dh.backend.clinicamvc.exception.BadRequestException;
+import dh.backend.clinicamvc.exception.ResourceNotFoundException;
 import dh.backend.clinicamvc.repository.IOdontologoRepository;
 import dh.backend.clinicamvc.service.IOdontologoService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +15,7 @@ import java.util.Optional;
 @Service
 public class OdontologoService implements IOdontologoService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(OdontologoService.class);
     private IOdontologoRepository odontologoRepository;
 
     public OdontologoService(IOdontologoRepository odontologoRepository) {
@@ -18,7 +23,11 @@ public class OdontologoService implements IOdontologoService {
     }
 
     @Override
-    public Odontologo registrarOdontologo(Odontologo odontologo) {
+    public Odontologo registrarOdontologo(Odontologo odontologo) throws BadRequestException {
+        if(odontologo.getNombre() == null || odontologo.getApellido() == null)
+            throw new BadRequestException("{\"message\": \"Error al crear el Odontólogo\"}");
+
+        LOGGER.info("Odontólogo persistido con exito");
         return odontologoRepository.save(odontologo);
     }
 
@@ -35,11 +44,18 @@ public class OdontologoService implements IOdontologoService {
     @Override
     public void actualizarOdontologo(Odontologo odontologo) {
         odontologoRepository.save(odontologo);
+        LOGGER.info("Odontólogo modificado con exito");
     }
 
     @Override
-    public void eliminarOdontologo(Integer id) {
-        odontologoRepository.deleteById(id);
+    public void eliminarOdontologo(Integer id) throws ResourceNotFoundException {
+        Optional<Odontologo> odontologoOptional = buscarPorId(id);
+        if (odontologoOptional.isPresent()) {
+            odontologoRepository.deleteById(id);
+            LOGGER.info("Odontólogo eliminado con exito");
+        } else {
+            throw new ResourceNotFoundException("{\"message\": \"Odontólogo no encontrado\"}");
+        }
     }
 
     @Override
