+++ BackEnd-I/Clinica-MVC/src/main/java/dh/backend/clinicamvc/service/impl/PacienteService.java
@@ -2,8 +2,12 @@ package dh.backend.clinicamvc.service.impl;
 
 
 import dh.backend.clinicamvc.entity.Paciente;
+import dh.backend.clinicamvc.exception.BadRequestException;
+import dh.backend.clinicamvc.exception.ResourceNotFoundException;
 import dh.backend.clinicamvc.repository.IPacienteRepository;
 import dh.backend.clinicamvc.service.IPacienteService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,6 +16,7 @@ import java.util.Optional;
 @Service
 public class PacienteService implements IPacienteService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(PacienteService.class);
     private IPacienteRepository pacienteRepository;
 
     public PacienteService(IPacienteRepository pacienteRepository) {
@@ -19,7 +24,11 @@ public class PacienteService implements IPacienteService {
     }
 
     @Override
-    public Paciente registrarPaciente(Paciente paciente) {
+    public Paciente registrarPaciente(Paciente paciente) throws BadRequestException {
+        if(paciente.getNombre() == null || paciente.getApellido() == null)
+            throw new BadRequestException("{\"message\": \"Error al crear el Paciente\"}");
+
+        LOGGER.info("Paciente persistido con exito");
         return pacienteRepository.save(paciente);
     }
 
@@ -36,11 +45,18 @@ public class PacienteService implements IPacienteService {
     @Override
     public void actualizarPaciente(Paciente paciente) {
         pacienteRepository.save(paciente);
+        LOGGER.info("Paciente modificado con exito");
     }
 
     @Override
-    public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+    public void eliminarPaciente(Integer id) throws ResourceNotFoundException {
+        Optional<Paciente> pacienteOptional = buscarPorId(id);
+        if(pacienteOptional.isPresent()) {
+            pacienteRepository.deleteById(id);
+            LOGGER.info("Paciente eliminado con exito");
+        } else {
+            throw new ResourceNotFoundException("{\"message\": \"Paciente no encontrado\"}");
+        }
     }
 
     @Override
